requirements.txt
streamlit>=1.28.0
ollama>=0.1.0
Pillow>=10.0.0
